apps/server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sku-lens-server"
version = "0.1.0"
description = "SKU Lens backend for AI Winner & Loser Analysis"
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
  "aiomysql>=0.2.0",
  "aiosqlite>=0.20.0",
  "cryptography>=44.0.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "orjson>=3.10.0",
  "pydantic-settings>=2.7.0",
  "redis>=6.0.0",
  "sqlalchemy>=2.0.38",
  "sqlmodel>=0.0.24",
  "uvicorn>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "asgi-lifespan>=2.1.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.26.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.11.0"
]

[project.scripts]
sku-lens-server = "main:run"
sku-lens-worker = "worker:main"

[tool.hatch.build.targets.wheel]
only-include = [
  "src/config.py",
  "src/db.py",
  "src/job_queue.py",
  "src/main.py",
  "src/models.py",
  "src/repositories",
  "src/schemas.py",
  "src/security",
  "src/services",
  "src/worker.py"
]
sources = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["ANN", "B", "E", "F", "I", "N", "S", "UP"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S106"]
